+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Change Log
 
+- python: getting history via api
 - nodejs: link in referenceData
 - nodejs: onRequestSelectDocument method
 - nodejs: onRequestSelectSpreadsheet method